Decoupled PostsIndex from action creator
PostsIndex is now wrapped in a custom Route implementation which handles the action creator which populates the object. This is better for reusability but this seems like a slightly clumsy implementation pattern. There is also some redux thunk code, but commented out for now since it does not work yet.